refactored guessTeamController, added dbGuessTeam to models
