+++ handlers/teams.js
@@ -1,5 +1,6 @@
 import query from "../dbConfig/index.js"
 import { dbGetTeams } from "../models/db/dbGetTeams.js";
+import { dbGuessTeam } from "../models/db/dbGuessTeam.js";
 import { Team } from "../models/requests/team.js";
 import { badRequest } from "../models/responses/badRequest.js";
 import { goodResponse200 } from "../models/responses/goodResponse.js";
@@ -17,15 +18,7 @@ export async function getTeamsController() {
 
 export async function guessTeamController() {
     try {
-        let teams = await query(`SELECT * FROM teams;`)
-        teams = teams.rows
-        const randNum = Math.floor(Math.random() * teams.length);
-        const randomTeam = teams[randNum]
-        const data = {
-            "teamid": randomTeam.teamid,
-            "anagram" : randomTeam.anagram
-            }
-        return goodResponse200(data)
+        return goodResponse200(await dbGuessTeam())
     } catch (err) {
         return badRequest(err.message)
     }
